travel_project/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-admin-sortable2==2.2.8
django-ckeditor==6.7.3
django-js-asset==3.1.2
flake8==7.3.0
mccabe==0.7.0
pillow==12.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
